Nest module layout.
